Split memcache client into generic and Frank parts
