+++ frankenstein/chunk_store_test.go
@@ -59,13 +59,13 @@ func TestChunkStore(t *testing.T) {
 	store := AWSChunkStore{
 		dynamodb:   mockaws.NewMockDynamoDB(),
 		s3:         mockaws.NewMockS3(),
-		memcache:   nil,
+		chunkCache: nil,
 		tableName:  "tablename",
 		bucketName: "bucketname",
 		cfg: ChunkStoreConfig{
-			S3URL:          "",
-			DynamoDBURL:    "",
-			MemcacheClient: nil,
+			S3URL:       "",
+			DynamoDBURL: "",
+			ChunkCache:  nil,
 		},
 	}
 	store.CreateTables()
